+++ DataRepo/views/models/__init__.py
@@ -1,6 +1,7 @@
 from .animal import AnimalDetailView, AnimalListView
 from .compound import CompoundDetailView, CompoundListView
 from .infusate import InfusateDetailView, InfusateListView
+from .lcmethod import LCMethodDetailView, LCMethodListView
 from .msrun import MSRunDetailView, MSRunListView
 from .peakdata import PeakDataListView
 from .peakgroup import PeakGroupDetailView, PeakGroupListView
@@ -19,6 +20,8 @@ __all__ = [
     "CompoundDetailView",
     "InfusateListView",
     "InfusateDetailView",
+    "LCMethodListView",
+    "LCMethodDetailView",
     "StudyListView",
     "StudyDetailView",
     "study_summary",