+++ changelog.md
@@ -13,11 +13,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added an ArchiveFile class with associated DataType and DataFormat classes
 - LCMethod model and tests
 - Copy records from PeakGroupSet to ArchiveFile, all records are copied as MS Peak Annotation data in AccuCor format as a default
+- LCMethod views
 
 ### Changed
 
 - Updated dependencies
 - Use human readable string methods for ArchiveFile, DataType, and DataFormat (affects admin interface, other default templates)
+- Added an LCMethod row/column to the MSRun views
 
 ## [2.0.3] - 2023-07-07